carefaq-rag/requirements.txt
streamlit
rank-bm25
scikit-learn
pytest

sentence-transformers
numpy
pyyaml
faiss-cpu
torch
